% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellAuthentication.R
\docType{package}
\name{CellAuthentication}
\alias{CellAuthentication}
\alias{CellAuthentication-package}
\title{CellAuthentication: a package for cell authentication.}
\description{
The CellAuthentication package provides a number of functions for performing
authentication and validation of biological samples through analysis of high
throughput sequencing data.
}
\section{Authentication functions}{

The Authentication functions ...
}

