% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_overlaps.R
\name{compare_overlaps}
\alias{compare_overlaps}
\title{Find matching variants in a dataframe of overlapping SNV profiles.}
\usage{
compare_overlaps(overlaps)
}
\arguments{
\item{overlaps}{The dataframe with the two overlapping SNV profiles}
}
\value{
A dataframe with an added 'match' column
}
\description{
This is a function for comparing the genotypes of each variant between two
overlapping SNV profiles contained in a single dataframe. A new column
('match') will be added to the dataframe, which indicates if each 
variant's genotype is a match or not (i.e. 'match' or 'mismatch').
}
\examples{
data(overlaps)
compare_overlaps(overlaps)
}
