% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_profiles.R
\name{compare_profiles}
\alias{compare_profiles}
\title{Overlap and compare genotypes in two SNV profiles.}
\usage{
compare_profiles(profile_1, profile_2)
}
\arguments{
\item{profile_1}{The first variant GRanges object.}

\item{profile_2}{The second variant GRanges object.}
}
\value{
A data frame.
}
\description{
This is a function for finding overlapping variants in two different SNV
profiles (stored as GenomicRanges objects), followed by comparing the
genotypes of the overlapping variants. The "compare_overlaps" function calls
the "add_metadata" function twice in succession in order to merge the
metadata for the two profiles (supplied as GRanges objects), returns the
results as a dataframe, compares the genotypes of the overlapping variants
using the "compare_genotypes" function and, finally, returns the final
dataframe with all variant overlaps and their similarity.
}
\examples{
data(test_profile_1)
data(test_profile_2)
compare_profiles(test_profile_1, test_profile_2)
}
