% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_variants.R
\name{compare_variants}
\alias{compare_variants}
\title{Find matching variants in a data frame.}
\usage{
compare_variants(data)
}
\arguments{
\item{overlaps}{The dataframe to analyse.}
}
\value{
A dataframe with an added 'match' column
}
\description{
This is a function for comparing matching or mismatching variants across
two samples contained in a single data frame.
}
\details{
A new column ('match') will be added to the data frame, which indicates if
each variant's genotype is a match or not (i.e. 'match' or 'mismatch').
}
\examples{
data(overlaps)
compare_variants(overlaps)
}
