% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile.R
\name{create_profile}
\alias{create_profile}
\title{Create an SNV profile from a VCF file}
\usage{
create_profile(vcf_file, sample, output_file, filter_depth = 10,
  python = FALSE)
}
\arguments{
\item{vcf_file}{The VCF file from which variants will be extracted.}

\item{sample}{The sample in the VCF that will be extracted.}

\item{output_file}{Results will be output to this file}

\item{filter_depth}{Remove variants below this sequencing depth}

\item{python}{Extract variants using Python instead of R}
}
\value{
Does not return any data object, but output results to output_file
}
\description{
This function creates a SNV profile from a given VCF file by extracting the
variants that pass the filtering criterias. It can either be performed using
R, or by the create_profile.py function included (which requires that Python
is installed, along with the PyVCF package). Profile creation is performed
to facilitate and accelerate the cell authentication procedures, which is
especially relevant when more than one pairwise comparison will be performed
on the same sample.
}
\examples{
vcf_file = system.file("extdata",
                       "example.vcf.gz", 
                       package = "CellAuthentication")
create_profile(vcf_file, "sample1", "profile1.txt")
create_profile(vcf_file, "sample1", "profile1.txt", filter_depth = 15)
create_profile(vcf_file, "sample1", "profile1.txt", python = TRUE)
}
