% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variants.R
\name{extract_variants}
\alias{extract_variants}
\title{Extract relevant variant information from a VCF file}
\usage{
extract_variants(vcf_file, sample)
}
\arguments{
\item{vcf_file}{The VCF file from which variants will be extracted.}

\item{sample}{The sample in the VCF that will be extracted.}

\item{output_file}{File to store results in, when not outputting to STDOUT.}
}
\value{
Returns the extracted variants to STDOUT or to a file.
}
\description{
This function is a wrapper to the "extract_variants.py" Python script, which
extracts relevant variant information from a VCF file. Extraction is
performed to facilitate and accelerate the cell authentication procedures,
which is especially relevant when more than one pairwise comparison will be
performed on the same sample.
}
\examples{
extract_variants(vcf_file, sample)
extract_variants_file(vcf_file, sample, output_file)
}
