% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_profiles.R
\name{overlap_profiles}
\alias{overlap_profiles}
\title{Find overlapping variants in two GenomicRanges objects.}
\usage{
overlap_profiles(profile_1, profile_2, sample_1 = "sample_1",
  sample_2 = "sample_2")
}
\arguments{
\item{profile_1}{The first variant GRanges object.}

\item{profile_2}{The second variant GRanges object.}

\item{sample_1}{Name of the first sample.}

\item{sample_2}{Name of the second sample.}
}
\value{
A data frame.
}
\description{
This is a function for finding overlapping variants in two different SNV
profiles (stored as GenomicRanges objects). The overlap_profiles function
is a wrapper function that adds metadata to an intersection of two GRanges
objects twice, in order to merge the metadata for those overlapping
profiles, returning a data frame.
}
\examples{
data(profile_1)
data(profile_2)
overlap_profiles(profile_1, profile_2, "sample1", "sample2")
}
