% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_variants.R
\name{overlap_variants}
\alias{overlap_variants}
\title{Find overlapping variants in two GenomicRanges objects.}
\usage{
overlap_variants(object_1, object_2, sample_1 = "sample_1",
  sample_2 = "sample_2")
}
\arguments{
\item{object_1}{The first variant GRanges object.}

\item{object_2}{The second variant GRanges object.}

\item{sample_1}{Name of the first sample.}

\item{sample_2}{Name of the second sample.}
}
\value{
A data frame.
}
\description{
This is a function for finding overlapping variants in two different
GenomicRanges objects.
}
\details{
The overlap_variants function is a wrapper function that adds metadata to an
intersection of two GRanges objects twice, in order to merge the metadata
for those overlapping variants, returning a data frame.
}
\examples{
data(granges_1)
data(granges_2)
overlap_variants(granges_1, granges_2, "sample1", "sample2")
}
