% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_profile.R
\name{read_profile}
\alias{read_profile}
\title{Read SNV profile}
\usage{
read_profile(file, sample_name)
}
\arguments{
\item{file}{The file path to be read.}

\item{sample_name}{The sample name from which the file originates.}
}
\value{
A GenomicRanges object.
}
\description{
\link{read_profile} reads SNV profiles as GRanges objects for
 downstream profile comparisons
}
\details{
This is a function for reading SNV profiles extracted from VCF
files.  The data is returned as a GenomicRanges object, suitable for merging
of metadata.
}
\examples{
# Path to test data
profile = system.file("extdata",
                      "test_profile_1.txt.gz",
                      package = "seqCAT")

# Read test profile
profile_1 <- read_profile(profile, "sample1")
}
