% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variants.R
\name{read_variants}
\alias{read_variants}
\title{Reads variant data in 'extracted' format.}
\usage{
read_variants(file, sample_name)
}
\arguments{
\item{file}{The file path to be read.}

\item{sample_name}{The sample name from which the file originates.}
}
\value{
A GenomicRanges object.
}
\description{
This is a function for reading variant data in the 'extracted' type format,
generated by the \code{extract_variants.py} script. The read data is
returned as a GenomicRanges object, suitable for merging of metadata.
}
\examples{
extract = system.file("extdata",
                      "extract.sample1.txt",
                      package = "CellAuthentication")
read_variants(extract, 'sample1')
}
