% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqCAT.R
\docType{package}
\name{seqCAT}
\alias{seqCAT}
\alias{seqCAT-package}
\title{seqCAT: High Throughput Sequencing Cell Authentication Toolkit}
\description{
The *seqCAT* package provides a number of functions for
performing evaluation, characterisation and authentication of biological
samples through analysis of high throughput sequencing data.
}
