% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_variant_list}
\alias{test_variant_list}
\title{Modified variant list object}
\format{A dataframe with 2 rows and 2 columns:
\describe{
     \item{sample1}{the genotypes of sample1}
     \item{sample2}{the genotypes of sample2}
}}
\usage{
data(test_variant_list)
}
\description{
A variant list object from the `list_variants` function, where the row names
have been defined as "chr: pos (gene)" and the corresponding columns
removed, for use in plotting.
}
\keyword{datasets}
