% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_matches.R
\name{variant_matches}
\alias{variant_matches}
\title{Find matching variants in a data frame.}
\usage{
variant_matches(data)
}
\arguments{
\item{data}{The dataframe to analyse.}
}
\value{
A dataframe with an added "match" column
}
\description{
This is a function for finding matching or mismatching variants across
two samples contained in a single data frame.
}
\details{
A new column ("match") will be added to the data frame, which indicates if
each variant is a perfect match or not (i.e. "match" or "mismatch").
}
\examples{
variant_matches(variants.df)
}
