% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variants_1}
\alias{variants_1}
\title{Example variants from "sample1"}
\format{A data frame with 383 rows and 20 columns:
\describe{
    \item{seqnames}{chromosome}
    \item{start}{start position}
    \item{end}{end position}
    \item{rsID}{mutation ID, if available}
    \item{gene}{associated gene}
    \item{ENSGID}{ensembl gene ID}
    \item{ENSTID}{ensembl transcript ID}
    \item{REF}{reference allele}
    \item{ALT}{alternative allele}
    \item{impact}{putative variant impact}
    \item{effect}{variant effect}
    \item{feature}{transcript feature}
    \item{biotype}{transcript biotype}
    \item{DP}{total variant depth}
    \item{AD1}{allelic depth, allele 1}
    \item{AD2}{allelic depth, allele 2}
    \item{A1}{allele 1}
    \item{A2}{allele 2}
    \item{warnings}{warnings from variant calling}
    \item{sample}{sample name}
}}
\usage{
data(variants_1)
}
\description{
Variants from "sample1" read with the read_variants function, for use in
examples and tests.
}
\keyword{datasets}
