% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_profile.R
\name{write_profile}
\alias{write_profile}
\title{Write SNV profile}
\usage{
write_profile(profile, file)
}
\arguments{
\item{profile}{The SNV profile to be written (data frame).}

\item{file}{The file to write to (path).}
}
\value{
None; writes to disk only.
}
\description{
Write an SNV profile to a file for later re-use.
}
\details{
This is a function for writing SNV profiles (created from VCF
files) to disk for later re-use. Several formats are allowed, including BED,
GTF, GFF and normal text files, which are automatically recognised based on
the supplied filename.
}
\examples{
# Load test profile
data(test_profile_1)

# Write test profile to file
write_profile(test_profile_1, "test_profile_1.txt")
}
