% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile.R
\name{create_profile}
\alias{create_profile}
\title{SNV profile creation}
\usage{
create_profile(vcf_file, sample, output_file, min_depth = 10,
  filter = TRUE)
}
\arguments{
\item{vcf_file}{The VCF file from which the profile will be created (path).}

\item{sample}{The sample in the VCF for which a profile will be created
(character).}

\item{output_file}{The output file with the SNV profile (path).}

\item{min_depth}{Remove variants below this sequencing depth (integer).}

\item{filter}{Remove variants not passing filtering criteria (boolean).}
}
\value{
Does not return any data object, but outputs results to output_file
 (to save computational time from having to repeatedly create profiles).
}
\description{
Create an SNV profile from data in a VCF file.
}
\details{
This function creates a SNV profile from a given VCF file by
extracting the variants that pass the filtering criterias. Profile creation
is performed to facilitate and accelerate the cell authentication
procedures, which is especially relevant when more than one pairwise
comparison will be performed on the same sample.
}
\examples{
# Path to the test VCF file
vcf_file = system.file("extdata", "test.vcf.gz", package = "seqCAT")

# Create SNV profiles
\dontrun{
 create_profile(vcf_file, "sample1", "profile1.txt")
 create_profile(vcf_file, "sample1", "profile1.txt", min_depth = 15)
}
}
