% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile.R
\name{create_profiles}
\alias{create_profiles}
\title{SNV profile creation}
\usage{
create_profiles(vcf_dir, output_dir = ".", pattern = NULL,
  recursive = FALSE, min_depth = 10, filter = TRUE, python = FALSE)
}
\arguments{
\item{vcf_dir}{The VCF directory from which the profiles will be created 
(path).}

\item{output_dir}{The output directory to put the SNV profiles in (path).}

\item{pattern}{Only create profiles for a subset of files corresponding to
this pattern (character).}

\item{recursive}{Find VCF files recursively in sub-directories as well
(boolean).}

\item{min_depth}{Remove variants below this sequencing depth (integer).}

\item{filter}{Remove variants not passing filtering criteria (boolean).}

\item{python}{Extract variants using Python instead of R (boolean).}
}
\value{
Does not return any data object, but outputs results to output_dir
 (to save computational time from having to repeatedly create profiles).
}
\description{
Create SNV profiles from all VCF files in a directory
}
\details{
This functions is a convenience-wrapper for the `create_profile`
function, which will create SNV profiles for each and every VCF file in the
provided directory. The file naming scheme used is `<sample>.vcf` and will
dictate the output profile filenames.
}
\examples{
# Path to the test VCF directory
vcf_dir = system.file("extdata", package = "seqCAT")

# Create SNV profiles
\dontrun{
 create_profiles(vcf_dir, output_dir = "profiles")
 create_profiles(vcf_dir, output_dir = "profiles", pattern = "test")
 create_profiles(vcf_dir, output_dir = "profiles", recursive = TRUE)
}
}
