% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variants.R
\name{extract_variants}
\alias{extract_variants}
\title{Extract relevant variant information from a VCF file}
\usage{
extract_variants(vcf_file, sample, output_file, filter_depth = 10,
  python = FALSE)
}
\arguments{
\item{vcf_file}{The VCF file from which variants will be extracted.}

\item{sample}{The sample in the VCF that will be extracted.}

\item{output_file}{Results will be output to this file}

\item{filter_depth}{Remove variants below this sequencing depth}

\item{python}{Extract variants using Python instead of R}
}
\value{
Does not return any data object, but output results to output_file
}
\description{
This function is a wrapper to the 'extract_variants.py' Python script, which
extracts relevant variant information from a VCF file. Extraction is
performed to facilitate and accelerate the cell authentication procedures,
which is especially relevant when more than one pairwise comparison will be
performed on the same sample.
}
\examples{
vcf_file = system.file("extdata",
                       "example.vcf.gz", 
                       package = "CellAuthentication")
sample = "sample1"
output_file = "test_extract.txt"
extract_variants(vcf_file, sample, output_file)
extract_variants(vcf_file, sample, output_file, filter_depth = 15)
extract_variants(vcf_file, sample, output_file, python = TRUE)
}
