% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants}
\alias{filter_variants}
\title{Filter variants on specified criteria}
\usage{
filter_variants(variants, overlaps = TRUE)
}
\arguments{
\item{variants}{The data frame containing the variant data to be filtered.}

\item{overlaps}{Boolean that governs filtering on complete genotypes
(enabled by default)}
}
\value{
A data frame containing the filtered variants.
}
\description{
This is a function for filtering variants on various criteria. Variants with
a depth lower than 10 are removed, as are variants that don't have a
complete genotype for both samples (by default).
}
\examples{
filter_variants(data)
filter_variants(data, overlaps = FALSE)
}
