% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants}
\alias{filter_variants}
\title{Filter variants on specified criteria}
\usage{
filter_variants(variants, filter_depth = 10)
}
\arguments{
\item{variants}{The data frame containing the variant data to be filtered.}

\item{filter_depth}{Threshold for variant depth (default 10)}
}
\value{
A data frame containing the filtered variants.
}
\description{
This is a function for filtering variants on sequencing depth. Variants with
a depth lower than 10 are removed by default, but can be changed in the
function call.
}
\examples{
filter_variants(data)
filter_variants(data, filter_depth=20)
}
