% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants}
\alias{filter_variants}
\title{Variant filtering}
\usage{
filter_variants(overlaps, filter_depth = 10)
}
\arguments{
\item{overlaps}{The data frame containing the variant data to be filtered.}

\item{filter_depth}{Threshold for variant depth (default 10)}
}
\value{
A data frame containing the filtered variants.
}
\description{
\link{filter_variants} filters variants on sequencing depth
}
\details{
This is a function for filtering variants on sequencing depth.
Variants with a depth lower than 10 are removed by default, but can be
changed in the function call.
}
\examples{
# Load test comparisons
data(test_comparison)

# Filter variants
filt_1 <- filter_variants(test_comparison)
filt_2 <- filter_variants(test_comparison, filter_depth = 20)
}
