% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cosmic.R
\name{list_cosmic}
\alias{list_cosmic}
\title{List COSMIC cell lines}
\usage{
list_cosmic(file_path)
}
\arguments{
\item{file_path}{The CosmicCLP_MutantExport.tsv.gz file (path).}
}
\value{
A vector of cell line names
}
\description{
List all available cell lines in the COSMIC database
}
\details{
This function lists the available cell lines in the provided
CosmicCLP_MutantExport.tsv.gz file, and takes about half the time it takes
to read the full file with the read_cosmic function, making it useful for
just seeing if your particular cell line is listed in COSMIC or not.
}
\examples{
file <- system.file("extdata",
                    "subset_CosmicCLP_MutantExport.tsv.gz",
                    package = "seqCAT")
cell_lines <- list_cosmic(file)
}
