% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_variants.R
\name{add_metadata}
\alias{add_metadata}
\alias{overlap_variants}
\title{Find overlapping variants in two GenomicRanges objects.}
\usage{
add_metadata(query, subject, column_suffix)

overlap_variants(object_1, object_2)
}
\arguments{
\item{query}{The query to add subject metadata to.}

\item{subject}{The subject whose metadata gets added to query.}

\item{column_suffix}{A string which will be added to all the metadata
columns from the current add_metadata operation}

\item{object_1}{The first variant GRanges object.}

\item{object_2}{The second variant GRanges object.}
}
\value{
Each function returns a single GRanges object.
}
\description{
This are functions for finding overlapping variants in two different
GenomicRanges objects, returning another GRanges object.
}
\details{
The add_metadata function is a function for adding metadata (i.e. any
column that is not the "seqnames", "start" or "end" fields in a
GenomicRanges object) from the "subject" GRanges object to the "query" 
GRanges object. The overlap_variants function, on the other hand, is a
wrapper function that calls add_metadata twice in succession, ensuring that
even non-overlapping variants are included in the final result.
}
\examples{
add_metadata(data_first, data_second)
overlap_variants(data_first, data_second)
}
