% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_variants.R
\name{add_metadata}
\alias{add_metadata}
\alias{overlap_variants}
\title{Find overlapping variants in two GenomicRanges objects.}
\usage{
add_metadata(query, subject, column_suffix)

overlap_variants(object_1, object_2, sample_1 = "sample_1",
  sample_2 = "sample_2")
}
\arguments{
\item{object_1}{The first variant GRanges object.}

\item{object_2}{The second variant GRanges object.}

\item{sample_1}{Name of the first sample.}

\item{sample_2}{Name of the second sample.}
}
\value{
A GRanges object.
}
\description{
This are functions for finding overlapping variants in two different
GenomicRanges objects, returning another GRanges object.
}
\details{
The add_metadata function is a function for adding metadata (i.e. any
column that is not the "seqnames", "start" or "end" fields in a
GenomicRanges object) from the "subject" GRanges object to the "query" 
GRanges object. The overlap_variants function, on the other hand, is a
wrapper function that calls add_metadata twice in succession, to add named
metadata to the intersection of variants between the two datasets.
}
\examples{
data(variants_1)
data(variants_2)
overlap_variants(variants_1, variants_2, "sample1", "sample2")
}
