% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot similarity heatmap}
\usage{
plot_heatmap(similarities, annotate = TRUE, annotate_size = 5,
  legend = TRUE, cluster = TRUE, limits = c(0, 50, 90, 100),
  colour = "#1954A6")
}
\arguments{
\item{similarities}{The long-format dataframe to be plotted}

\item{annotate}{Annotate each cell with the score}

\item{annotate_size}{The size of the annotations}

\item{legend}{Show a legend for the colour gradient}

\item{cluster}{Cluster the samples based on similarity}

\item{limits}{The limits for the colour gradient}

\item{colour}{The main colour to use for the gradient}
}
\value{
A ggplot2 graphical object
}
\description{
\link{plot_heatmap} plots a heatmap of similarities from 
 multiple sample comparisons
}
\details{
This function creates publication-ready plots of heatmaps for
many-to-many  sample comparisons, taking a long-format dataframe containing
the summary statistics of each comparison as input.
}
\examples{
# Load test similarities
data(test_similarities)

# Plot a similarity heatmap
heatmap <- plot_heatmap(test_similarities)
}
