% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impacts.R
\name{plot_impacts}
\alias{plot_impacts}
\title{Plot SNV impact distributions for a binary comparison}
\usage{
plot_impacts(comparison, annotate = TRUE, legend = TRUE,
  palette = c("#0D2D59", "#1954A6"))
}
\arguments{
\item{comparison}{The profile comparison to be plotted}
}
\description{
This function creates publication-ready plots of the impact distribution
from a binary dataset comparison across the matched/mismatched SNVs.
}
\examples{
data(test_overlaps)
plot_impacts(test_overlaps)
}
