% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cosmic.R
\name{read_cosmic}
\alias{read_cosmic}
\alias{list_cosmic}
\title{Read COSMIC cell line-specific mutational data}
\usage{
read_cosmic(file_path, cell_line)

list_cosmic(file_path)
}
\arguments{
\item{file_path}{Location of the CosmicCLP_MutantExport.tsv.gz file}

\item{cell_line}{The cell line to be investigated}

\item{file_path}{Location of the CosmicCLP_MutantExport.tsv.gz file}
}
\value{
A GRanges object with COSMIC SNVs
}
\description{
This function reads the "CosmicCLP_MutantExport.tsv.gz" file obtained from
COSMIC and returns a GRanges object with all the listed mutations for the
specified cell line, which can then be use in downstream profile
comparisons. Only non-duplicated (gene-level) SNVs are included in COSMIC
profiles.

This function lists the available cell lines in the provided
CosmicCLP_MutantExport.tsv.gz file, and take about half the time it takes to
read the full file with the read_cosmic function, making it useful for just
seeing if your particular cell line is listed in COSMIC or not.
}
\examples{
file <- system.file("extdata",
                    "subset_CosmicCLP_MutantExport.tsv.gz",
                    package = "CellAuthentication")
cosmic_hct116 <- read_cosmic(file, "HCT116")
file <- system.file("extdata",
                    "subset_CosmicCLP_MutantExport.tsv.gz",
                    package = "CellAuthentication")
cell_lines <- list_cosmic(file)
}
