% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_profile.R
\name{read_profile}
\alias{read_profile}
\title{Read SNV profile}
\usage{
read_profile(file, sample_name)
}
\arguments{
\item{file}{The SNV profile to be read (path).}

\item{sample_name}{The sample of the SNV profile (character).}
}
\value{
A GRanges object.
}
\description{
Read SNV profiles for use in downstream comparisons.
}
\details{
This is a function for reading SNV profiles created from VCF
files.  The data is returned as a GenomicRanges object, suitable for merging
of metadata.
}
\examples{
# Path to test data
profile = system.file("extdata",
                      "test_profile_1.txt.gz",
                      package = "seqCAT")

# Read test profile
profile_1 <- read_profile(profile, "sample1")
}
