% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_profile.R
\name{read_profile}
\alias{read_profile}
\title{Read SNV profile}
\usage{
read_profile(file, sample_name, remove_mt = TRUE)
}
\arguments{
\item{file}{The SNV profile to be read (path).}

\item{sample_name}{The sample of the SNV profile (character).}

\item{remove_mt}{Remove or keep mitochondrial variants (boolean).}
}
\value{
A GRanges object.
}
\description{
Read an SNV profile for use in downstream comparisons.
}
\details{
This is a function for reading SNV profiles created from VCF
files.  The data is returned as a GenomicRanges object, suitable for merging
of metadata.
}
\examples{
# Path to test data
profile = system.file("extdata",
                      "test_1.profile.txt.gz",
                      package = "seqCAT")

# Read test profile
profile_1 <- read_profile(profile, "sample1")
profile_1 <- read_profile(profile, "sample1", remove_mt = FALSE)
}
