% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_profile.R
\name{read_profile}
\alias{read_profile}
\title{Reads SNV profiles}
\usage{
read_profile(file, sample_name)
}
\arguments{
\item{file}{The file path to be read.}

\item{sample_name}{The sample name from which the file originates.}
}
\value{
A GenomicRanges object.
}
\description{
This is a function for reading SNV profiles extracted from VCF files.
The data is returned as a GenomicRanges object, suitable for merging of
metadata.
}
\examples{
profile = system.file("extdata",
                      "test_profile_1.txt.gz",
                      package = "CellAuthentication")
read_profile(profile, "sample1")
}
